feat: Add async submission to `llm_generate` (#1965)
* Create an internal PhoenixException class that can be caught

* Extract executors into their own module

* Add tests and fix sync executor retry logic

* Add tests for retry behavior

* Implement llm_generate using executors

* Lazily initialize async primitives in rate limiter

* Manage termination event lifecycle in executors

* Revert to using `get_executor_on_sync_context`

* Add typing information

* Improve sigint handling test

* Update tests for executor refactor

* Use correct fallback values for `llm_generate`

* Use SIGUSR1 to test signal handling instead of SIGINT

* Address PR feedback
