+++ src/phoenix/exceptions.py
@@ -1,0 +1,2 @@
+class PhoenixException(Exception):
+    pass
